+++ lib/ruby_aem_aws/component/publish.rb
@@ -62,6 +62,15 @@ module RubyAemAws
         instance.terminate
         instance.wait_until_terminated
       end
+
+      def get_tags
+        tags = []
+        get_all_instances.each do |i|
+          next if i.nil?
+          tags.push(i.tags)
+        end
+        tags
+      end
     end
   end
 end